Added completed Project 1 and 2 files
Project 1 - Predicting Boston Housing Prices

The task is to practice:
- initially exploring data
- training and testing models
- identifying overfitting/underfitting issues
- applying cross-validation and tuning techniques such as GridSearchCV and Kfold.

Project 2 - Supervised Learning Project - Finding Donors for CharityML

The task is to explore census data and attempt to predict whether a resident earns >50K or <=50K annually. This information can then be used to target donation brochures for CharityML, a fictional charity organization.

The task explores and compares various supervised learning algorithms. An optimal algorithm is ultimately chosen and tuned.